Move crawlers to subpackage, add abstract base and group support
