+++ dirstree/__init__.py
@@ -1,2 +1,2 @@
-from dirstree.crawler import Crawler as Crawler  # noqa: F401
-from dirstree.python_crawler import PythonCrawler as PythonCrawler  # noqa: F401
+from dirstree.crawlers.crawler import Crawler as Crawler  # noqa: F401
+from dirstree.crawlers.python_crawler import PythonCrawler as PythonCrawler  # noqa: F401
